API CHANGE: Add a new class to sysadm: sysadm/firewall This is the new firewall manager (ipfw), for setting options for the system firewall.
Initial API Call: "action":"known_ports"
This will return a list of all known ports and any names/descriptions for them (this is a static list - it does not reflect which ports are in-use or opened on the system. It is just for matching a port to a name/description)

REST Request (example):
-------------------------------
PUT /sysadm/firewall
{
   "action" : "known_ports"
}

WebSocket Request:
-------------------------------
{
   "id" : "fooid",
   "namespace" : "sysadm",
   "args" : {
      "action" : "known_ports"
   },
   "name" : "firewall"
}

Response:
-------------------------------
{
  "args": {
    "1/tcp": {
      "description": "#TCP Port Service Multiplexer",
      "name": "tcpmux",
      "port": "1/tcp"
    },
    "1/udp": {
      "description": "#TCP Port Service Multiplexer",
      "name": "tcpmux",
      "port": "1/udp"
    },
    "100/tcp": {
      "description": "#[unauthorized use]",
      "name": "newacct",
      "port": "100/tcp"
    }
  }
  "id": "fooid",
  "name": "response",
  "namespace": "sysadm"
}

+++ src/server/library/sysadm-firewall.cpp
@@ -93,7 +93,7 @@ QList<PortInfo> Firewall::allPorts(){
     PortInfo info;
     info.Keyword = line[0];
     info.Port = line[1].section("/",0,0).toInt();
-    info.Type = line[1].section("/",1,1).toInt();
+    info.Type = line[1].section("/",1,1);
     if(line.length()>2){
       info.Description = line.mid(2,-1).join(" ");
     }
